requirements.txt
async-generator==1.10
attrs==22.1.0
autopep8==1.7.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.3.0
certifi==2023.7.22
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorzero==2.0
distlib==0.3.6
exceptiongroup==1.0.0rc9
filelock==3.9.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mypy-extensions==0.4.3
outcome==1.2.0
packaging==23.0
paho-mqtt==1.6.1
pathspec==0.10.1
platformdirs==2.6.2
pluggy==1.0.0
pycodestyle==2.9.1
pyproject_api==1.5.0
PySocks==1.7.1
pytest==7.2.1
python-dotenv==0.21.0
requests==2.31.0
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
tox==4.4.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.17.1
webdriver-manager==3.8.3
wsproto==1.2.0
